BackEnd/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Flask-APScheduler
psycopg2-binary
python-dotenv
Flask-CORS